🌈 Initial commit of an HTTP request function with timeout support
The design goal for this package is to provide a fetch-like HTTP
request function that supports a configurable timeout, because the
native `fetch()` timeout is too short.

Currently, this implementation proxies to `XMLHttpRequest` to satisfy
requests. Soon, it should have alternative implementations based on
the features available at runtime and/or a configuration option where
a compatible implementation is injected or specified.